+++ packages/drivers/odsp-driver/src/contracts.ts
@@ -4,7 +4,7 @@
  */
 
 import * as api from "@fluidframework/protocol-definitions";
-import { IOdspUrlParts, HostStoragePolicy } from "@fluidframework/odsp-driver-definitions";
+import { HostStoragePolicy } from "@fluidframework/odsp-driver-definitions";
 
 /**
  * Socket storage discovery api response
@@ -182,31 +182,4 @@ export interface ICreateFileResponse {
     itemId: string;
     itemUrl: string;
     sequenceNumber: number;
-}
-
-/**
- * @deprecated - use OdspFluidDataStoreLocator
- */
-export type OdspDocumentInfo = OdspFluidDataStoreLocator;
-
-export interface OdspFluidDataStoreLocator extends IOdspUrlParts {
-    dataStorePath: string;
-    appName?: string;
-    containerPackageName?: string;
-    fileVersion?: string;
-}
-
-export enum SharingLinkHeader {
-    // Can be used in request made to resolver, to tell the resolver that the passed in URL is a sharing link
-    // which can be redeemed at server to get permissions.
-    isSharingLinkToRedeem = "isSharingLinkToRedeem",
-}
-
-export interface ISharingLinkHeader {
-    [SharingLinkHeader.isSharingLinkToRedeem]: boolean;
-}
-
-declare module "@fluidframework/core-interfaces" {
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface
-    export interface IRequestHeader extends Partial<ISharingLinkHeader> { }
 }
