+++ packages/drivers/odsp-driver/src/createOdspUrl.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import {  OdspFluidDataStoreLocator } from "./contracts";
+import {  OdspFluidDataStoreLocator } from "./contractsPublic";
 
 /*
  * Per https://github.com/microsoft/FluidFramework/issues/1556, isolating createOdspUrl() in its own file.
